+++ lib/osm/section.rb
@@ -1,59 +1,74 @@
 module Osm
 
   class Section
+    class FlexiRecord; end # Ensure the constant exists for the validators
 
-    attr_reader :id, :name, :subscription_level, :subscription_expires, :type, :num_scouts, :column_names, :fields, :intouch_fields, :mobile_fields, :flexi_records, :role
-    # @!attribute [r] id
+    include ::ActiveAttr::MassAssignmentSecurity
+    include ::ActiveAttr::Model
+
+    # @!attribute [rw] id
     #   @return [Fixnum] the id for the section
-    # @!attribute [r] name
+    # @!attribute [rw] name
     #   @return [String] the section name
-    # @!attribute [r] subscription_level
+    # @!attribute [rw] subscription_level
     #   @return [Symbol] what subscription the section has to OSM (:bronze, :silver or :gold)
-    # @!attribute [r] subscription_expires
+    # @!attribute [rw] subscription_expires
     #   @return [Date] when the section's subscription to OSM expires
-    # @!attribute [r] type
+    # @!attribute [rw] type
     #   @return [Symbol] the section type (:beavers, :cubs, :scouts, :exporers, :adults, :waiting, :unknown)
-    # @!attribute [r] num_scouts
+    # @!attribute [rw] num_scouts
     #   @return [Fixnum] how many members the section has
-    # @!attribute [r] column_names
+    # @!attribute [rw] column_names
     #   @return [Hash] custom names to use for the data columns
-    # @!attribute [r] fields
+    # @!attribute [rw] fields
     #   @return [Hash] which columns are shown in OSM
-    # @!attribute [r] intouch_fields
+    # @!attribute [rw] intouch_fields
     #   @return [Hash] which columns are shown in OSM's in touch reports
-    # @!attribute [r] mobile_fields
+    # @!attribute [rw] mobile_fields
     #   @return [Hash] which columns are shown in the OSM mobile app
-    # @!attribute [r] flexi_records
+    # @!attribute [rw] flexi_records
     #   @return [Array<FlexiRecord>] list of the extra records the section has
-    # @!attribute [r] role
+    # @!attribute [rw] role
     #   @return [Osm::Role] the role linking the user to this section
 
+    attribute :id, :type => Integer
+    attribute :name, :type => String
+    attribute :subscription_level, :default => :unknown
+    attribute :subscription_expires, :type => Date
+    attribute :type, :default => :unknown
+    attribute :num_scouts, :type => Integer
+    attribute :column_names, :default => {}
+    attribute :fields, :default => {}
+    attribute :intouch_fields, :default => {}
+    attribute :mobile_fields, :default => {}
+    attribute :flexi_records, :default => []
+    attribute :role
 
-    # Initialize a new Section
-    # @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
-    def initialize(attributes={})
-      raise ArgumentError, ':id must be nil or a Fixnum > 0' unless attributes[:id].nil? || (attributes[:id].is_a?(Fixnum) && attributes[:id] > 0)
-      raise ArgumentError, ':section_name must be nil or a String' unless attributes[:section_name].nil? || attributes[:section_name].is_a?(String)
-      raise ArgumentError, ':num_scouts must be nil or a Fixnum >= 0' unless attributes[:num_scouts].nil? || (attributes[:num_scouts].is_a?(Fixnum) && attributes[:num_scouts] >= 0)
-      [:column_names, :fields, :intouch_fields, :mobile_fields].each do |attribute|
-        raise ArgumentError, ":#{attribute} must be nil or a Hash" unless attributes[attribute].nil? || attributes[attribute].is_a?(Hash)
-      end
-      [:type, :subscription_level].each do |attribute|
-        raise ArgumentError, ":#{attribute} must be nil or a Symbol" unless attributes[attribute].nil? || attributes[attribute].is_a?(Symbol)
-      end
-      raise ArgumentError, ':flexi_records must be nil or an Array' unless attributes[:flexi_records].nil? || attributes[:flexi_records].is_a?(Array)
+    attr_accessible :id, :name, :subscription_level, :subscription_expires, :type, :num_scouts, :column_names, :fields, :intouch_fields, :mobile_fields, :flexi_records, :role
 
-      attributes.each { |k,v| instance_variable_set("@#{k}", v) }
+    validates_numericality_of :id, :only_integer=>true, :greater_than=>0, :allow_nil => true
+    validates_numericality_of :num_scouts, :only_integer=>true, :greater_than_or_equal_to=>0
+    validates_presence_of :subscription_level
+    validates_presence_of :subscription_expires
+    validates_presence_of :type
+    validates_presence_of :column_names, :unless => Proc.new { |a| a.column_names == {} }
+    validates_presence_of :fields, :unless => Proc.new { |a| a.fields == {} }
+    validates_presence_of :intouch_fields, :unless => Proc.new { |a| a.intouch_fields == {} }
+    validates_presence_of :mobile_fields, :unless => Proc.new { |a| a.mobile_fields == {} }
+    validates_presence_of :flexi_records, :unless => Proc.new { |a| a.flexi_records == [] }
+    validates_presence_of :role
 
-      @section_name ||= ''
-      @column_names ||= {}
-      @fields ||= {}
-      @intouch_fields ||= {}
-      @mobile_fields ||= {}
-      @flexi_records ||= []
-      @subscription_level ||= :unknown
-      @type ||= :unknown
-    end
+    validates_inclusion_of :subscription_level, :in => [:bronze, :silver, :gold, :unknown], :message => 'is not a valid level'
+
+    validates :column_names, :hash => {:key_type => Symbol, :value_type => String}
+    validates :fields, :hash => {:key_type => Symbol, :value_type => String}
+    validates :intouch_fields, :hash => {:key_type => Symbol, :value_type => String}
+    validates :mobile_fields, :hash => {:key_type => Symbol, :value_type => String}
+
+
+    # @!method initialize
+    #   Initialize a new Section
+    #   @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
 
 
     # Initialize a new Sections from api data
@@ -80,12 +95,10 @@ module Osm
         :flexi_records => [],
       }
 
-
       # Populate arrays
       (data['extraRecords'].is_a?(Array) ? data['extraRecords'] : []).each do |record_data|
         attributes[:flexi_records].push FlexiRecord.from_api(record_data)
       end
-      attributes[:flexi_records].freeze
 
       new(attributes)
     end
@@ -93,7 +106,7 @@ module Osm
     # Check if this section is one of the youth sections
     # @return [Boolean]
     def youth_section?
-      [:beavers, :cubs, :scouts, :explorers].include?(@type)
+      [:beavers, :cubs, :scouts, :explorers].include?(type)
     end
 
     # Custom section type checkers
@@ -117,7 +130,7 @@ module Osm
     #   @return (Boolean)
     [:beavers, :cubs, :scouts, :explorers, :adults, :waiting].each do |attribute|
       define_method "#{attribute}?" do
-        @type == attribute
+        type == attribute
       end
     end
 
@@ -136,29 +149,49 @@ module Osm
         return false
       end
     end
+    
+    def inspect
+      attribute_descriptions = attributes.merge('role' => role.inspect_without_section(self))
+      return_inspect(attribute_descriptions)
+    end
 
+    def inspect_without_role(exclude_role)
+      attribute_descriptions = (role == exclude_role) ? attributes.merge('role' => 'SET') : attributes
+      return_inspect(attribute_descriptions)
+    end
 
 
     private
-    class FlexiRecord
+    def return_inspect(attribute_descriptions)
+      attribute_descriptions.sort.map { |key, value| "#{key}: #{key.eql?('role') ? value : value.inspect}" }.join(", ")
+      separator = " " unless attribute_descriptions.empty?
+      "#<#{self.class.name}#{separator}#{attribute_descriptions}>"
+    end
 
-      attr_reader :id, :name
-      # @!attribute [r] id
+    class FlexiRecord
+      include ::ActiveAttr::MassAssignmentSecurity
+      include ::ActiveAttr::Model
+
+      # @!attribute [rw] id
       #   @return [Fixnum] the aid of the flexi-record
-      # @!attribute [r] name
+      # @!attribute [rw] name
       #   @return [String] the name given to the flexi-record
-  
-      # Initialize a new ApiAccess
-      # @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
-      def initialize(attributes={})
-        raise ArgumentError, ':id must be a Fixnum > 0' unless (attributes[:id].is_a?(Fixnum) && attributes[:id] > 0)
-        raise ArgumentError, ':name must be a String' unless attributes[:name].is_a?(String)
-  
-        attributes.each { |k,v| instance_variable_set("@#{k}", v) }
-      end
-  
-  
-      # Initialize a new ApiAccess from api data
+
+      attribute :id, :type => Integer
+      attribute :name, :type => String
+
+      attr_accessible :id, :name
+
+      validates_numericality_of :id, :only_integer=>true, :greater_than=>0
+      validates_presence_of :name
+
+
+      # @!method initialize
+      #   Initialize a new FlexiRecord
+      #   @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
+
+
+      # Initialize a new FlexiRecord from api data
       # @param [Hash] data the hash of data provided by the API
       def self.from_api(data)
         # Expect item to be: {:name=>String, :extraid=>Fixnum}
@@ -170,8 +203,8 @@ module Osm
           :name => data['name'],
         })
       end
-    end # FlexiRecord
+    end # Class Section::FlexiRecord
 
-  end # Section
+  end # Class Section
 
 end # Module