+++ spec/osm/evening_spec.rb
@@ -6,7 +6,7 @@ describe "Evening" do
 
   before :each do
     @attributes = {
-      :evening_id => 1,
+      :id => 1,
       :section_id => 2,
       :title => 'Evening Name',
       :notes_for_parents => 'Notes for parents',
@@ -15,7 +15,7 @@ describe "Evening" do
       :post_notes => 'After',
       :leaders => 'Leaders',
       :start_time => '19:00',
-      :end_time => '21:00',
+      :finish_time => '21:00',
       :meeting_date => Date.new(2000, 01, 02),
     }
   end
@@ -42,7 +42,7 @@ describe "Evening" do
     }]
     e = Osm::Evening.from_api(data, activities)
 
-    e.evening_id.should == 1
+    e.id.should == 1
     e.section_id.should == 2
     e.title.should == 'Evening Name'
     e.notes_for_parents.should == 'Notes for parents'
@@ -51,30 +51,15 @@ describe "Evening" do
     e.post_notes.should == 'After'
     e.leaders.should == 'Leaders'
     e.start_time.should == '19:00'
-    e.end_time.should == '21:00'
+    e.finish_time.should == '21:00'
     e.meeting_date.should == Date.new(2000, 1, 2)
 
     ea = e.activities[0]
     ea.activity_id.should == 2
     ea.title.should == 'Activity Name'
     ea.notes.should == 'Notes'
-  end
 
-
-  it "Raises exceptions when trying to set invalid times" do
-    e = Osm::Evening.new(@attributes)
-    
-    expect{ e.start_time = 'abcde' }.to raise_error(ArgumentError)
-    expect{ e.start_time = '24:00' }.to raise_error(ArgumentError)
-    expect{ e.start_time = '10:61' }.to raise_error(ArgumentError)
-    e.start_time = '12:34'
-    e.start_time.should == '12:34'
-
-    expect{ e.end_time = 'abcde' }.to raise_error(ArgumentError)
-    expect{ e.end_time = '24:00' }.to raise_error(ArgumentError)
-    expect{ e.end_time = '10:61' }.to raise_error(ArgumentError)
-    e.end_time = '23:45'
-    e.end_time.should == '23:45'
+    e.valid?.should be_true
   end