+++ spec/spec_helper.rb
@@ -1,6 +1,8 @@
 require 'fakeweb'
 require 'httparty'
 require 'active_support'
+require 'active_attr/rspec'
+require 'active_model'
 
 require 'osm'