+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## Version 0.0.18
+
+  * Term's end attribute is now finish
+  * Event's end attribute is now finish
+  * Evening's evening\_id attribute is now id
+  * DueBadge's totals attribute is now a method (totals are calculated each time rather than during initialization)
+  * Added exception ArgumentIsInvalid (raised when argument.valid? is false when updating through the API)
+  * The following models now use active\_attr:
+    * Activity, Activity::File, Activity::Badge and Activity::Version
+    * ApiAccess
+    * DueBadges
+    * Evening and Evening::Activity
+    * Event
+    * Grouping
+    * Member
+    * RegisterData
+    * RegisterField
+    * Role
+    * Section and Section::FlexiRecord
+    * Term
+
 ## Version 0.0.17
 
   * Fix try method is undefined